feat(GoogleMapsAPILoader): change default API version
This changes the default version of the Google Maps API to `quarterly`
because the Google Maps team changed the version strategy.

BREAKING CHANGES
This could change the Google Maps API version - so please check if you
need to change the setting of the `apiVersion` attribute in the
`AgmCodeModule.forRoot()` call.