feat: Add placeId to map click events
* feat: allow disabling a "point of interest" default info window
* feat: Add placeId to map click events

when clicking on a point of interest on a map the emiting event will include the `placeId` of the location clicked.
allow disabling a "point of interest" default info window. Use it by passing a boolean value to the new property isToShowDefaultInfoWindowForIcons when initializing the map. Default value is `true`.

closes issue #691.
closes issue #1539.