feat: remove AgmMouseEvent in favor of google's
In line with removing custom types, remove also the MouseEvent type
which converted LatLng to LatLngLiteral.
Previously, behavior was inconsistent, with some directives emitting
the custom MouseEvent, and others claiming to do so, but in fact
emitting the original MouseEvent.
Now all Mouse events will emit the original google's MouseEvent or
IconMouseEvent.